+++ packages/git/src/index.ts
@@ -11,6 +11,7 @@
 export {
 	BranchEntry,
 	CommitInfo,
+	CommitLogEntry,
 	ConfigListEntry,
 	DirtyWorktreeError,
 	Git,