+++ packages/tsconfig-json/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @effected/tsconfig-json
 
+## 0.8.1
+
+### Dependencies
+
+| Dependency | Type | Action | From | To |
+| --- | --- | --- | --- | --- |
+| @effected/walker | dependency | updated | 0.6.0 | 0.7.0 |
+
 ## 0.8.0
 
 ### Breaking Changes